requirements.txt
Django>=4.1.7,<4.2
django-filter>=22.1,<22.2
django-cors-headers>=3.14.0,<3.15
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.26.0,<0.27
psycopg2-binary>=2.9.6,<3.0
Pillow>=9.4.0,<9.5
redis>=4.5.2,<4.6
celery>=5.2.7,<5.3
requests>=2.28.2,<2.29
PyJWT>=2.6.0,<2.7
channels>=4.0.0,<4.1
django-celery-beat>=2.5.0,<2.6
channels-redis>=4.1.0,<4.2